Refactored run.py. Constants were placed in constants.py Improved error handling. Added gcp integration
