+++ control_tower/cloud/__init__.py
@@ -1,0 +1,2 @@
+from .aws import create_aws_instances
+from .gcp import create_gcp_instances
